More toInt changes

+++ hail/src/main/scala/is/hail/annotations/Region.scala
@@ -93,15 +93,15 @@ object Region {
   }
 
   def setBit(byteOff: Long, bitOff: Long) {
-    val b = byteOff + (bitOff >> 3)
+    val b = byteOff.toInt + (bitOff.toInt >> 3)
     storeByte(b,
-      (loadByte(b) | (1 << (bitOff & 7))).toByte)
+      (loadByte(b) | (1 << (bitOff.toInt & 7))).toByte)
   }
 
   def clearBit(byteOff: Long, bitOff: Long) {
-    val b = byteOff + (bitOff >> 3)
+    val b = byteOff.toInt + (bitOff.toInt >> 3)
     storeByte(b,
-      (loadByte(b) & ~(1 << (bitOff & 7))).toByte)
+      (loadByte(b) & ~(1 << (bitOff.toInt & 7))).toByte)
   }
 
   def storeBit(byteOff: Long, bitOff: Long, b: Boolean) {
